+++ packages/files-ui/src/Components/Modules/FileBrowsers/views/FileSystemItem/FileSystemGridItem.tsx
@@ -15,6 +15,7 @@ import { ConnectDragPreview } from "react-dnd"
 import { Form, FormikProvider, useFormik } from "formik"
 import { nameValidator } from "../../../../../Utils/validationSchema"
 import Menu from "../../../../../UI-components/Menu"
+import { getFileNameAndExtension } from "../../../../../Utils/Helpers"
 
 const useStyles = makeStyles(({ breakpoints, constants, palette }: CSFTheme) => {
   return createStyles({
@@ -169,26 +170,7 @@ const FileSystemGridItem = React.forwardRef(
     const [isEditingLoading, setIsEditingLoading] = useState(false)
 
     const { fileName, extension } = useMemo(() => {
-      if (isFolder) {
-        return {
-          fileName : name,
-          extension: ""
-        }
-      }
-      const split = name.split(".")
-      const extension = `.${split[split.length - 1]}`
-
-      if (split.length === 1) {
-        return {
-          fileName : name,
-          extension: ""
-        }
-      }
-
-      return {
-        fileName: name.slice(0, name.length - extension.length),
-        extension: split[split.length - 1]
-      }
+      return getFileNameAndExtension(name, isFolder)
     }, [name, isFolder])
 
     const formik = useFormik({
