+++ packages/files-ui/src/Components/Modules/FileBrowsers/views/FileSystemItem/FileSystemTableItem.tsx
@@ -20,6 +20,7 @@ import { ConnectDragPreview } from "react-dnd"
 import { Form, FormikProvider, useFormik } from "formik"
 import { nameValidator } from "../../../../../Utils/validationSchema"
 import Menu from "../../../../../UI-components/Menu"
+import { getFileNameAndExtension } from "../../../../../Utils/Helpers"
 
 const useStyles = makeStyles(({ breakpoints, constants, palette }: CSFTheme) => {
   const desktopGridSettings = "50px 69px 3fr 190px 100px 45px !important"
@@ -155,26 +156,7 @@ const FileSystemTableItem = React.forwardRef(
     const [isEditingLoading, setIsEditingLoading] = useState(false)
 
     const { fileName, extension } = useMemo(() => {
-      if (isFolder) {
-        return {
-          fileName : name,
-          extension: ""
-        }
-      }
-      const split = name.split(".")
-      const extension = `.${split[split.length - 1]}`
-
-      if (split.length === 1) {
-        return {
-          fileName : name,
-          extension: ""
-        }
-      }
-
-      return {
-        fileName: name.slice(0, name.length - extension.length),
-        extension: split[split.length - 1]
-      }
+      return getFileNameAndExtension(name, isFolder)
     }, [name, isFolder])
 
     const formik = useFormik({
